kube-test: script fixes and improvements
Type: test

+++ test-c/kube-test/script/quick-import.sh
@@ -4,8 +4,13 @@ CALICOVPP_DIR=${CALICOVPP_DIR:-"$HOME/vpp-dataplane"}
 VPP_DIR=$(pwd)
 VPP_DIR=${VPP_DIR%test-c*}
 
+# Tag of built CalicoVPP images.
+# CALICOVPP_VERSION should be the same as TAG when running kube-test
+TAG=${TAG:-"kt-master"}
+BASE=${BASE:-"origin/master"}
+
 if [ "$1" = "" ]; then
-    echo "This script will save and import images to both nodes.
+    echo "This script will build, save and import images to both nodes.
 To import kube-test images only, \$2 = kt
 To import CalicoVPP images only, \$2 = cv
 To import all, leave \$2 empty.
@@ -30,19 +35,32 @@ fi
 
 if [ "$2" = "cv" ] || [ "$2" = "" ]; then
     if [ ! -d "$CALICOVPP_DIR" ]; then
-          git clone https://github.com/projectcalico/vpp-dataplane.git $CALICOVPP_DIR
+      git clone https://github.com/projectcalico/vpp-dataplane.git $CALICOVPP_DIR
     else
-        echo "Repo found, resetting"
-        cd $CALICOVPP_DIR
-        git reset --hard origin/master
-        git pull
+      echo "Repo found, resetting"
+      cd $CALICOVPP_DIR
+      git reset --hard origin/master
+      git fetch --tags --force
+      git pull
+      cd vpp-manager
+      rm vpp*.tar || true
+      make clean-vpp
+      if [[ -d "$CALICOVPP_DIR/vpp-manager/vpp_build" ]]; then
         cd $CALICOVPP_DIR/vpp-manager/vpp_build
+        make wipe || true
+        make wipe-release || true
         git reset --hard origin/master
-        cd $VPP_DIR/test-c/kube-test
-    fi
+        git fetch --tags --force
+        git pull
+      fi
 
-    make -C $CALICOVPP_DIR image TAG=latest
-    docker save -o calicovpp-images.tar docker.io/calicovpp/vpp:latest docker.io/calicovpp/agent:latest docker.io/calicovpp/multinet-monitor:latest
+      cd $VPP_DIR/test-c/kube-test
+  fi
+
+	make -C $CALICOVPP_DIR/vpp-manager vpp BASE=$BASE
+    make -C $CALICOVPP_DIR dev TAG=$TAG
+    make -C $CALICOVPP_DIR image TAG=$TAG
+    docker save -o calicovpp-images.tar docker.io/calicovpp/vpp:$TAG docker.io/calicovpp/agent:$TAG docker.io/calicovpp/multinet-monitor:$TAG
     sudo ctr -n k8s.io images import calicovpp-images.tar
     scp calicovpp-images.tar $1
     ssh $remote_user "sudo ctr -n k8s.io images import \"$remote_path\""/calicovpp-images.tar
